Move swagger_ui mod to a separate crate
