+++ Cargo.toml
@@ -3,6 +3,7 @@ members = [
     "okapi",
     "rocket-okapi",
     "rocket-okapi-codegen",
+    "rocket-okapi-ui",
     "examples/json-web-api",
     "examples/secure_request_guard"
 ]
