+++ examples/secure_request_guard/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2018"
 schemars = { version = "0.8"}
 okapi = { version = "0.6.1", path = "../../okapi", package = "okapi_fork"}
 rocket_okapi = { version = "0.8.0-rc.1", path = "../../rocket-okapi", package = "rocket_okapi_fork"}
+rocket_okapi_ui = { version = "0.1.0-rc.1", path = "../../rocket-okapi-ui" }
 
 [dependencies.rocket]
 git              = "https://github.com/SergioBenitez/Rocket"
