+++ rocket-okapi/src/handlers/mod.rs
@@ -1,5 +1,3 @@
 mod openapi;
-mod redirect;
 
 pub use openapi::*;
-pub use redirect::*;
